benchmark/redisgraph/requirements.txt
Click>=7.0
redis>=3.3.8
requests>=2.22.0
hdrhistogram
